Refactor enemy classes

+++ src/main/java/com/sentinel/starbasesentinel/levels/Grid.java
@@ -6,7 +6,7 @@ public class Grid {
     // 1280x720 is converted to 26x15 grid
     final static int cellSize = 48; // Cell size in pixels
 
-    public static Point2D translate(int x, int y) {
+    public static Point2D translate(double x, double y) {
         return new Point2D(x * cellSize + 24, y * cellSize + 24);
     }
 }
